Backend/API/api481W/requirements.txt
flask
fastapi
mysql-connector-python
